requirements.txt

streamlit
pandas
openpyxl
python-docx
